Lift and shift extraction to the Opcodes module
