+++ core-strato/ethereum-vm/src/Blockchain/VM/Opcodes.hs
@@ -1,15 +1,27 @@
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE MagicHash #-}
+{-# LANGUAGE BangPatterns #-}
 module Blockchain.VM.Opcodes where
 
 import           Prelude                      hiding (EQ, GT, LT)
 
-import           Data.Binary
+import           Data.Bits
 import qualified Data.ByteString              as B
+import qualified Data.ByteString.Unsafe       as BU
 import           Data.Data
+import           Data.Primitive.ByteArray
+import           Foreign.Ptr
+import           Foreign.Storable
+import           GHC.Exts
+import           GHC.Integer.GMP.Internals (Integer(..), BigNat(..))
+import           GHC.Word
+import           System.Endian
+import           System.IO.Unsafe
 import qualified Data.Map                     as M
 import           Data.Maybe
 import           Text.PrettyPrint.ANSI.Leijen hiding ((<$>))
 
+import           Network.Haskoin.Internals (BigWord(..), Word256)
 import           Blockchain.Util
 
 type CodePointer = Int
@@ -208,4 +220,49 @@ opCode2Op rom =
              $ M.lookup opcode code2OpMap in
     (op, 1)
 
+-- Unoptimized extraction, for 8-24 bytes that are too infrequently seen
+-- to bother writing a specialization.
+defaultExtract :: B.ByteString -> Int -> Int -> Word256
+-- TODO(tim): Use fastBytesToWord256 once available
+defaultExtract bs off len = fromIntegral
+                                 . bytes2Integer
+                                 . B.unpack
+                                 . B.take len
+                                 . B.drop off
+                                 $ bs
 
+-- Used to push 1 byte
+fastExtractByte :: B.ByteString-> Int -> Word256
+fastExtractByte !code !off = let !(W8# b#) = BU.unsafeIndex code off
+                             in BigWord (S# (word2Int# b#))
+
+-- Used to push 2-7 bytes
+fastExtractSingle :: B.ByteString-> Int -> Int -> Word256
+fastExtractSingle !code !off !len = unsafePerformIO . BU.unsafeUseAsCString code $ \ptr -> do
+  let !offPtr = castPtr ptr :: Ptr Word64
+      !delta = 64 - (8 * len)
+  -- This may read past the end of the bytestring, but if the read is allowed
+  -- those garbage bytes are truncated by the shift.
+  !rawBits <- peekByteOff offPtr off
+  let !(W64# w#) = toBE64 rawBits `shiftR` delta
+  return $! BigWord (S# (word2Int# w#))
+
+-- Used to push 25-32 bytes
+fastExtractQuad :: B.ByteString -> Int -> Int -> Word256
+fastExtractQuad !code !off !len = unsafePerformIO . BU.unsafeUseAsCString code $ \ptr -> do
+  let !offPtr = castPtr (plusPtr ptr (off + len)) :: Ptr Word64
+  !dst <- newByteArray 32
+  !ll <- peekElemOff offPtr (-1)
+  !lh <- peekElemOff offPtr (-2)
+  !hl <- peekElemOff offPtr (-3)
+  -- This might be a violation: we read before the beginning of the bytestring.
+  -- However if the read is allowed, the garbage bytes are masked off.
+  !hh <- peekElemOff offPtr (-4)
+
+  writeByteArray dst 0 $! toBE64 ll
+  writeByteArray dst 1 $! toBE64 lh
+  writeByteArray dst 2 $! toBE64 hl
+  let !mask = bit (8 * (len - 24)) - 1
+  writeByteArray dst 3 $! toBE64 hh .&. mask
+  !(ByteArray ba#) <- unsafeFreezeByteArray dst
+  return (BigWord (Jp# (BN# ba#)))
